Fix small warning (signed-unsigned comparison).

+++ aslam_cv_common/include/aslam/common/eigen-hash.h
@@ -12,7 +12,7 @@ struct hash<Eigen::Matrix<Scalar, Rows, Cols>> {
   // https://wjngkoh.wordpress.com/2015/03/04/c-hash-function-for-eigen-matrix-and-vector/
   size_t operator()(const Eigen::Matrix<Scalar, Rows, Cols>& matrix) const {
     size_t seed = 0;
-    for (size_t i = 0; i < matrix.size(); ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(matrix.size()); ++i) {
       Scalar elem = *(matrix.data() + i);
       seed ^=
           std::hash<Scalar>()(elem) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
